crispr-ambiguous-mapping/pyproject.toml
[tool.poetry]
name = "crispr-ambiguous-mapping"
version = "0.0.131"
description = ""
authors = ["Basheer Becerra <bbecerr@outlook.com>"]
readme = "README.md"
packages = [{include = "crispr_ambiguous_mapping"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
numpy = "^1.24.4"
scipy = "^1.10.1"
matplotlib = "^3.7.1"
biopython = "^1.81"
pandarallel = "^1.6.4"
typeguard = "^3.0.2"
pandas = "^1.5.3"
seaborn = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
